Restructure to support Babel
