+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["env"]
+  "presets": ["env", "stage-0"]
 }
