+++ .gitignore
@@ -1,1 +1,2 @@
-config/  
+config/
+dist
