+++ source/index.js
@@ -1,5 +1,3 @@
-require('babel-register');
-
 import express from 'express';
 import mongoose from 'mongoose';
 import bodyParser from 'body-parser';
